tlabel/requirements.txt
django==2.1.3
cached-property==1.5.1
python-telegram-bot[socks]==11.1.0
pyopenssl==18.0.0
requests==2.20.0
psycopg2-binary==2.7.5